move reaction utils to cobra.core.Reaction
